Fixed multiple log file creation on startup
Added the createOnDemand property for each of the Log4j2 appenders so that empty files aren't created, also changed the log filename pattern to not include seconds, because multiple sets of all log files would be created for each second in the first few seconds on app startup. Also changed the argline in the pom to use the correct log4j2-test.yml file.